Move workflow setup specific endpoints to their own router
