+++ src/murfey/server/demo_api.py
@@ -38,6 +38,12 @@ from murfey.server import shutdown as _shutdown
 from murfey.server import templates
 from murfey.server.api import MurfeySessionID
 from murfey.server.api.auth import validate_token
+from murfey.server.api.session_info import Visit
+from murfey.server.api.workflow import (
+    DCGroupParameters,
+    DCParameters,
+    ProcessingJobParameters,
+)
 from murfey.server.murfey_db import murfey_db
 from murfey.util.config import MachineConfig, from_file, security_from_file
 from murfey.util.db import (
@@ -62,14 +68,11 @@ from murfey.util.db import (
 from murfey.util.models import (
     ClientInfo,
     CurrentGainRef,
-    DCGroupParameters,
-    DCParameters,
     FoilHoleParameters,
     FractionationParameters,
     GainReference,
     GridSquareParameters,
     PostInfo,
-    ProcessingJobParameters,
     ProcessingParametersSPA,
     ProcessingParametersTomo,
     RegistrationMessage,
@@ -82,7 +85,6 @@ from murfey.util.models import (
     TiltSeriesGroupInfo,
     TiltSeriesInfo,
     TomoProcessFile,
-    Visit,
 )
 from murfey.util.processing_params import default_spa_parameters
 from murfey.workflows.spa.picking import _register_picked_particles_use_diameter
