+++ src/murfey/server/main.py
@@ -18,9 +18,10 @@ import murfey.server.api.display
 import murfey.server.api.hub
 import murfey.server.api.instrument
 import murfey.server.api.processing_parameters
+import murfey.server.api.session_info
 import murfey.server.api.spa
+import murfey.server.api.workflow
 import murfey.server.websocket
-import murfey.util.models
 from murfey.server import template_files
 from murfey.util.config import get_security_config
 
@@ -76,6 +77,10 @@ app.include_router(murfey.server.api.display.router)
 app.include_router(murfey.server.api.instrument.router)
 app.include_router(murfey.server.api.hub.router)
 app.include_router(murfey.server.api.processing_parameters.router)
+app.include_router(murfey.server.api.session_info.router)
+app.include_router(murfey.server.api.session_info.spa_router)
+app.include_router(murfey.server.api.session_info.tomo_router)
+app.include_router(murfey.server.api.workflow.router)
 app.include_router(murfey.server.websocket.ws)
 
 # Search external packages for additional routers to include in Murfey
